Moved modules around to better group diff modules, and prepare for adding new commands later
