+++ vessel/diff/helpers/unified_diff.py
@@ -32,7 +32,7 @@ from typing import Any, Optional
 
 import portion  # type: ignore
 
-from vessel.utils.flag import Flag
+from vessel.diff.helpers.flag import Flag
 
 logger = getLogger(__name__)
